+++ examples/hbd_example/hbd_example.ino
@@ -9,7 +9,7 @@ const int E = Q / 2;
 const int S = Q / 4;
 
 Piezzo piezzo (piezzopin);
-double scale [] = {G5,Q, G5,Q, A5,Q, G5,Q,  C5,Q,  B5,Q,  R , Q, G5,Q,  G5,Q, 
+float scale [] = {G5,Q, G5,Q, A5,Q, G5,Q,  C5,Q,  B5,Q,  R , Q, G5,Q,  G5,Q, 
 A5, Q, G5, Q, D5, Q, C5, Q, R,Q,  G5,Q,  G5,Q,  G5, Q, E5,Q,  C5, Q, B5,Q, 
 A5,Q, F5,Q,  F5, Q, E5, Q, C5,Q,  D5,Q,  C5,Q, END};
